caps/les.md: add les/2 protocol specification (#64)
The specification was adapted from the documents in the
zsfelfoldi/go-ethereum wiki. The documents about BloomBits, the CHT and
flow control are merged into the specification preceding the list of
protocol messages. Message definitions are a flat list of messages in
the current protocol version les/2 with changes from les/1 listed in the
change log.

The texts are slightly edited to make the document easier to read,
but no technical changes were made to the spec.